reorganize utils a bit
