+++ tests/test_hf_gpt2_serialize.py
@@ -13,8 +13,8 @@ from utils import skip_if_no_torch
 
 import haliax as hax
 from levanter.config import TrainerConfig
-from levanter.modeling_utils import cross_entropy_loss
 from levanter.models.gpt2 import Gpt2LMHeadModel
+from levanter.utils.modeling_utils import cross_entropy_loss
 
 
 @skip_if_no_torch
